Remove custom profile fields
