+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_07_15_211923) do
+ActiveRecord::Schema.define(version: 2021_07_20_042422) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "citext"
@@ -427,17 +427,6 @@ ActiveRecord::Schema.define(version: 2021_07_15_211923) do
     t.bigint "user_id"
     t.index ["purchase_id", "purchase_type"], name: "index_credits_on_purchase_id_and_purchase_type"
     t.index ["spent"], name: "index_credits_on_spent"
-  end
-
-  create_table "custom_profile_fields", force: :cascade do |t|
-    t.string "attribute_name", null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.string "description"
-    t.citext "label", null: false
-    t.bigint "profile_id", null: false
-    t.datetime "updated_at", precision: 6, null: false
-    t.index ["label", "profile_id"], name: "index_custom_profile_fields_on_label_and_profile_id", unique: true
-    t.index ["profile_id"], name: "index_custom_profile_fields_on_profile_id"
   end
 
   create_table "data_update_scripts", force: :cascade do |t|
@@ -1518,7 +1507,6 @@ ActiveRecord::Schema.define(version: 2021_07_15_211923) do
   add_foreign_key "comments", "users", on_delete: :cascade
   add_foreign_key "credits", "organizations", on_delete: :restrict
   add_foreign_key "credits", "users", on_delete: :cascade
-  add_foreign_key "custom_profile_fields", "profiles", on_delete: :cascade
   add_foreign_key "devices", "consumer_apps"
   add_foreign_key "devices", "users"
   add_foreign_key "discussion_locks", "articles"
